(#24) Seperate ActionList from the GameSquare
Generation of action list didn't belong in the Model in the first place. The View will now render according to if the player is allowed to perform actions.

This logic is extracted to the view as this is specific to how the view wants to render this.

extracted action determination from the model

need to delete the code and remove the code from the model

unit tests are blocking me from doing it right now, will resolve it in the next commit

bug fix for saber in versus mode to be able to move

id and player was not set properly in versus scene